@@ -1,22 +1,44 @@
-import { Component } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { AuthentificationService } from '../services/authentification.service';
+import { Component, inject } from '@angular/core';
+import { AuthentificationService } from '../services/authentificationService/authentification.service';
+import { FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatInputModule } from '@angular/material/input';
+import { MatRadioModule } from '@angular/material/radio';
+import { MatSelectModule } from '@angular/material/select';
 
 @Component({
   selector: 'app-login',
   standalone: true,
-  imports: [FormsModule],
+  imports: [
+    MatInputModule,
+    MatButtonModule,
+    MatSelectModule,
+    MatRadioModule,
+    MatCardModule,
+    ReactiveFormsModule,
+    MatDialogModule,
+  ],
   templateUrl: './login.component.html',
-  styleUrl: './login.component.css'
+  styleUrl: './login.component.scss',
 })
 export class LoginComponent {
-   password: string = '';
-   username: string = '';
- constructor(private authService: AuthentificationService) {
+  private fb = inject(FormBuilder);
+  loginForm = this.fb.group({
+    email: [null, Validators.required],
+    password: [null, Validators.required],
+  });
+  constructor(private authService: AuthentificationService) {}
 
- }
-
- onLogin() {
-    this.authService.authenticate(this.username, this.password);
- }
+  onLogin() {
+    if (this.loginForm.valid) {
+      if (this.loginForm.value.email && this.loginForm.value.password) {
+        this.authService.authenticate(
+          this.loginForm.value.email,
+          this.loginForm.value.password
+        );
+      }
+    }
+  }
 }
